📜 comment: 주석 추가

+++ backend/src/main/java/org/example/backend/settlement/dto/response/SettlementDetailResponse.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Getter;
 import org.example.backend.settlement.entity.SettlementDetail;
 
+import java.math.BigDecimal;
+
 // 정산서 상세 내역 Dto
 
 @Getter
@@ -11,9 +13,9 @@ import org.example.backend.settlement.entity.SettlementDetail;
 public class SettlementDetailResponse {
     private Long id;
     private String orderName;      // 상품명 (스냅샷)
-    private String sourceType;     // "일반 상품", "캔디 후원" (한글)
+    private String sourceType;     // "일반 상품", "캔디"
     private Long salesAmount;      // 판매 금액
-    private double shareRatio;     // 적용된 비율 (0.9 or 0.2)
+    private BigDecimal shareRatio;     // 적용된 비율 (0.9 or 0.2)
     private Long settlementAmount; // 최종 인정액
 
     public static SettlementDetailResponse from(SettlementDetail entity) {
@@ -22,7 +24,7 @@ public class SettlementDetailResponse {
                 .orderName(entity.getTitleSnapshot())
                 .sourceType(entity.getSourceType().getDescription()) // Enum 설명 활용
                 .salesAmount(entity.getSalesAmount())
-                .shareRatio(entity.getShareRatio().doubleValue())
+                .shareRatio(entity.getShareRatio())
                 .settlementAmount(entity.getSettlementAmount())
                 .build();
     }
